slice k & attn mask fp16

+++ python/llm/src/ipex_llm/transformers/npu_models/qwen2_mp.py
@@ -143,8 +143,10 @@ class LowBitQwenMultiDecoderlayer(LLMBaseNNFactory):
             attention_mask = self.create_input_op(
                 (self.batch_size, 1, 1, self.max_seq_len + 1), dtype=np.int64)
         else:
+            # attention_mask = self.create_input_op(
+            #     (self.batch_size, 1, self.seq_len, self.seq_len), dtype=np.int64)
             attention_mask = self.create_input_op(
-                (self.batch_size, 1, self.seq_len, self.seq_len), dtype=np.int64)
+                (self.batch_size, 1, self.seq_len, self.seq_len), dtype=np.float16)
 
         position_ids = self.create_input_op((self.batch_size, self.seq_len), dtype=np.int64)
 
@@ -533,7 +535,8 @@ class FusedQwenLowBitDecoderlayer(torch.nn.Module):
 
         backend_cls = self.backend_cls_prefill
         inputs = (hidden_states.to(torch.float16),
-                  attention_mask.to(torch.int64),
+                #   attention_mask.to(torch.int64),
+                  attention_mask.to(torch.float16),
                   position_ids.to(torch.int64))
         inputs += (self.layer_norm_0, self.layer_norm_1, self.q_bias)
         hidden_states, past_key, past_value = run_model(
